fix: deprecate quest.service
