+++ tests/utils/questService.test.js
@@ -1,4 +1,4 @@
-import { fetchQuestCategoryData } from '@services/questService';
+import { fetchQuestCategoryData } from "@services/apiService";
 
 const API_URL = process.env.NEXT_PUBLIC_API_LINK;
